requirements.txt
faiss-cpu 
sentence-transformers
requests
BeautifulSoup
flask 
openai 
google-cloud-translate 
google-cloud-texttospeech 
SpeechRecognition 
pyttsx3
CORS
python-dotenv
flask-cors
groq




